After four years, I realized that the original implementation of atItem::compare was stupid (yes, I wrote it, and it was stupid).  Here's why:
The original purpose of the compare method was to be able to provide a total
ordering to be able to use items as key values in associative containers
(atMap, for example).  The original default compare was based on the type
names of the classes involved, which is completely useless when associative
containers are involved (all key objects will be the same class, and as a
result, will compare as identical).  This gets even more confusing when you
consider that the default equals() method is based on pointer values (objects
will compare as identical, but will not be equal).

In cases where a total ordering is obvious (atString, for example), the
class can and should provide a compare() method, which allows an
equivalent instance of the desired key value to be used to access the
container.  For example, if object ABC is added to a map using an atString
with the value "ABC", it can be accessed later by a completely different
string instance with the same "ABC" value.  This is good.  However, there
are other cases where we actually want to map a particular object to another
object, directly by instance.  One example is the mapping of vsWindowSystems
in VESS, to their associated vsWindow.  We're never going to try to create
another "equivalent" vsWindowSystem to look up that vsWindow, so it should
be OK for the map to be able to use vsWindowSystems as a key, and it's
perfectly reasonable for the total ordering of those keys to be based on
the object's address.

For these reasons, I changed the default atItem compare method to compare
based on address instead of type name.