Fixing shared memory allocation in Windows
Summary:
Unlike Linux, shared memory keys are not relinquished in Windows until
all users detached from the memory.  When attempting to allocate more space
with an existing key, it would simply return the handle to the existing shared
memory without resizing it.  To avoid this, the key given to Windows was turned
into a sufficiently unique tuple in order to not clash with existing shared
memory.

Test Plan:
Tested with AFS through atSharedMemoryInterface by sending 10000+
byte data packets through atSharedQueue.

Reviewers: martin, szielinski

Reviewed By: martin

Differential Revision: http://cr.irl.ucf.edu/D31